resolve issues with ec2
