+++ workflows/environment_terminate.go
@@ -16,6 +16,7 @@ func NewEnvironmentTerminator(ctx *common.Context, environmentName string) Execu
 		workflow.environmentDbTerminator(environmentName, ctx.StackManager, ctx.StackManager, ctx.StackManager),
 		workflow.environmentEcsTerminator(environmentName, ctx.StackManager, ctx.StackManager),
 		workflow.environmentConsulTerminator(environmentName, ctx.StackManager, ctx.StackManager),
+		workflow.environmentElbTerminator(environmentName, ctx.StackManager, ctx.StackManager),
 		workflow.environmentVpcTerminator(environmentName, ctx.StackManager, ctx.StackManager),
 	)
 }
@@ -95,6 +96,23 @@ func (workflow *environmentWorkflow) environmentEcsTerminator(environmentName st
 	return func() error {
 		log.Noticef("Terminating ECS environment '%s' ...", environmentName)
 		envStackName := common.CreateStackName(common.StackTypeCluster, environmentName)
+		err := stackDeleter.DeleteStack(envStackName)
+		if err != nil {
+			return err
+		}
+
+		stack := stackWaiter.AwaitFinalStatus(envStackName)
+		if stack != nil && !strings.HasSuffix(stack.Status, "_COMPLETE") {
+			return fmt.Errorf("Ended in failed status %s %s", stack.Status, stack.StatusReason)
+		}
+
+		return nil
+	}
+}
+func (workflow *environmentWorkflow) environmentElbTerminator(environmentName string, stackDeleter common.StackDeleter, stackWaiter common.StackWaiter) Executor {
+	return func() error {
+		log.Noticef("Terminating ELB environment '%s' ...", environmentName)
+		envStackName := common.CreateStackName(common.StackTypeLoadBalancer, environmentName)
 		err := stackDeleter.DeleteStack(envStackName)
 		if err != nil {
 			return err
